+++ include/network.h
@@ -15,10 +15,10 @@ struct Network {
   void init_network(unsigned int *inputs, unsigned int *neurons,
                     unsigned int clayers);
 
-  void propagate_network(const float *input);
+  void propagate_network(const double *input);
 
-  float train_network(const float *input, const float *awaited_output,
-                      const float learning_rate, float momentum);
+  double train_network(const double *input, const double *awaited_output,
+                       const double learning_rate, double momentum);
 
   ~Network();
 };
